Closes #220
