+++ setup.cfg
@@ -39,7 +39,7 @@ per-file-ignores =
   # We need `assert`s to please mypy:
   classes/contrib/mypy/*.py: S101
   # There are multiple assert's in tests:
-  tests/*.py: S101, WPS226, WPS432, WPS436
+  tests/*.py: S101, WPS226, WPS431, WPS432, WPS436
 
 
 [isort]
