+++ typesafety/test_typeclass/test_callback.yml
@@ -13,7 +13,6 @@
     def _example_int_float(instance: Union[int, float], attr: bool) -> bool:
         ...
 
-
     def accepts_typeclass(callback: Callable[[int, bool], bool]) -> bool:
         return callback(1, True)
 
@@ -38,8 +37,7 @@
     def accepts_typeclass(callback: Callable[[str, bool], bool]) -> bool:
         return callback('a', True)
 
-    reveal_type(accepts_typeclass(example))
+    accepts_typeclass(example)
   out: |
     main:16: error: Argument 1 to "accepts_typeclass" has incompatible type "_TypeClass[float, Callable[[Any, bool], bool], <nothing>, Literal['main.example']]"; expected "Callable[[str, bool], bool]"
     main:16: note: "_TypeClass[float, Callable[[Any, bool], bool], <nothing>, Literal['main.example']].__call__" has type "Callable[[Arg(Union[float, Supports[<nothing>]], 'instance'), VarArg(Any), KwArg(Any)], _ReturnType]"
-    main:16: note: Revealed type is 'builtins.bool'
